requirements.txt
pypng=="0.20220715.0"